(UNTESTED) Update file diff tool on server side
This updates a few files relating to the file diff enhancement. The
update is NOT UNIT TESTED and may not work at all. Logically, it should
though. The file diffing works according to the following:

* Obtains two strings for contents and URL
* Queries db with URL for filename
* Queries fs to obtain file contents
* Runs diff, returns the object containing an array of changes.

This can be worked with a post request from the client side.

Note that nothing has been done on the client side, so as of now it is
unknown if it works for real at all. Unit testing is necessary. Will
post an update later.

~ From somewhere in the middle of the Arabian ocean, altitude ~30000ft

Signed-off-by: Siddarth Iyer <mindstormer619@gmail.com>