(FIXED) Update Unit Tested filediff function
Filediff functionality is working as expected along with Unit Testing.

Just click the "Diff Test!" button to run the test on the current cont-
ents of the editor along with the previous saved version [0]. The resul-
ts are saved in a object named 'diffObj' which is returned by jsdiff
module. This can only be run on an object with one or  more than one
prev file versions. For a list of prev versions, check variable
`p_test`.

Note that the dev console must stay open for running unit test.

Signed-off-by: Siddarth Iyer <mindstormer619@gmail.com>